feat: adiciona exceções customizadas, tratamento de exceções, resposta de erro e validação com Bean Validation

+++ src/main/java/edu/douglaslima/dscommerce/service/ProductService.java
@@ -1,14 +1,19 @@
 package edu.douglaslima.dscommerce.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import edu.douglaslima.dscommerce.dto.ProductDTO;
 import edu.douglaslima.dscommerce.entity.Product;
 import edu.douglaslima.dscommerce.repository.ProductRepository;
+import edu.douglaslima.dscommerce.service.exception.DatabaseException;
+import edu.douglaslima.dscommerce.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,7 +23,8 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = productRepository.findById(id).get();
+		Product product = productRepository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado!"));
 		return new ProductDTO(product);
 	}
 	
@@ -37,15 +43,26 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO productDTO) {
-		Product product = productRepository.getReferenceById(id);
-		productDTO.copyToEntity(product);
-		product = productRepository.save(product);
-		return new ProductDTO(product);
+		try {
+			Product product = productRepository.getReferenceById(id);
+			productDTO.copyToEntity(product);
+			product = productRepository.save(product);
+			return new ProductDTO(product);
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);;
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+		try {
+			productRepository.deleteById(id);;
+		} catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Erro de integridade referencial!");
+		}
 	}
 
 }
